// src/types/consumedItem.ts
// Generated by https://quicktype.io

import { FoodItem } from "./foodItem";

export interface consumedItem {
  id?: number;
  food_id: number | null;
  name: string;
  description: string | null;
  kcal: number | null;
  water: number | null;
  protein: number | null;
  fat: number | null;
  carbohydrate: number | null;
  fiber: number | null;
  alcohol: number | null;
  grams: number;
  created_at?: string;
  updated_at?: string;
  foodItem?: FoodItem;
}